+++ packages/twenty-server/src/engine/metadata-modules/ai/ai-agent-execution/utils/mapDBPartsToUIMessageParts.ts
@@ -15,6 +15,7 @@ export const mapDBPartToUIMessagePart = (
       return {
         type: 'reasoning',
         text: part.reasoningContent ?? '',
+        state: (part.state as 'streaming' | 'done') ?? 'done',
       };
     case 'file':
       return {
